Tweak pb_reverse_lookup
and fix earlier bug I introduced, I was failing to convert a list back
to a string with str.join. The `"".join` thing looks bizarre when you
first see and it's certainly one of the ugliest parts of python imo. You
get used to it though; this is the standard way of converting from list
to string.