Use argparse rather than handrolled parser.
The argparse module is what the Python standard library provides for
parsing command line arguments. It has it's warts but it gets the job
done and it's pretty much the way most people do it. There are some
newer, fancier libraries, but if you want to stick to the standard
library, argparse is the way to go; it'll provide better error handling
and a nicer interface than hand-rolled code.

Note that there's  more refactoring that could be done here, this
was just the least invasive changes necessary to get a working version
with argparse. See future commits . . .