let argparse do handler dispatch
argparse has a feature where you can associate a function to call with
each subparser and then dispatch to that function based on the
command. Here I just cleaned up your handlers and added a few more to
make things consistent and then let argparse do it's thing instead of
having a big branching if statement.